So this commit:

- does *not* attempt to add any new annotations
- does *not* attempt to preserve existing annotations on non-visible
  (e.g., `private`) APIs, though it does frequently do so when those
  annotations present no merge conflicts

**Also**, this commit does **not** keep some annotations on `equals`
parameters in null-unmarked code. I didn't _set out_ to do that, but
given the choice between resolving ~140 _more_ merge conflicts and _not_
doing that, I chose "not." The annotation matters essentially not at all
because Kotlin already lets callers pass `null` (because the parameter
has unspecified nullness), and the Checker Framework already has stubs
where necessary. We should still think about adding those back, but it
will be trivial to do when we want it, and we can at least skip classes
that aren't `public` this time.

Note also that I did not try to restore `ICC_Profile`, which we'd
deleted in
https://github.com/jspecify/jdk/commit/0d2d6b6e5515313d48478d39115e689cd13a3767
after previously encountering problems with it.

I kept the union of the `.gitignore` entries from both repos.

I notice that this commit adds some workflow. It would be neat for us to
adapt those to actually build our annotated JDK, thereby catching at
least simple mistakes. Or maybe we should just delete them. For now, I'm
leaving them in place, even though they fail because the files refer to
JSpecify annotations, which don't exist in this repo.

+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/code/VMRegImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2006, 2017, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2006, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -30,6 +30,8 @@ import sun.jvm.hotspot.debugger.*;
 import sun.jvm.hotspot.runtime.*;
 import sun.jvm.hotspot.types.*;
 import sun.jvm.hotspot.utilities.*;
+import sun.jvm.hotspot.utilities.Observable;
+import sun.jvm.hotspot.utilities.Observer;
 
 public class VMRegImpl {
 
@@ -51,7 +53,7 @@ public class VMRegImpl {
     Type type = db.lookupType("VMRegImpl");
     AddressField stack0Field = type.getAddressField("stack0");
     stack0Addr = stack0Field.getValue();
-    stack0Val = (int) stack0Addr.hashCode();
+    stack0Val = stack0Addr.hashCode();
     stack0 = new VMReg(stack0Val);
     regNameField = type.getAddressField("regName[0]");
     stackSlotSize = db.lookupIntConstant("VMRegImpl::stack_slot_size");
