+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/opto/Block.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -30,6 +30,8 @@ import sun.jvm.hotspot.debugger.*;
 import sun.jvm.hotspot.runtime.*;
 import sun.jvm.hotspot.oops.*;
 import sun.jvm.hotspot.types.*;
+import sun.jvm.hotspot.utilities.Observable;
+import sun.jvm.hotspot.utilities.Observer;
 
 public class Block extends VMObject {
   static {
@@ -68,7 +70,7 @@ public class Block extends VMObject {
   }
 
   public double freq() {
-    return (double)freqField.getValue(getAddress());
+    return freqField.getValue(getAddress());
   }
 
   public Node_List nodes() {
